+++ project/src/const.ts
@@ -1,7 +1,5 @@
 const INITIAL_GENRE = 'All genres';
 
-const INITIAL_FILM_ID = 0;
-
 const AppRoute = {
   Main: '/',
   SignIn: '/login',
@@ -65,7 +63,6 @@ const FilmInfo = {
 
 export {
   INITIAL_GENRE,
-  INITIAL_FILM_ID,
   AppRoute,
   APIRoute,
   NameSpace,
