+++ project/src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import {Provider} from 'react-redux';
 import App from './components/app/app';
-import {store} from './store';
+import {store} from './store/store';
 import {
   checkAuthAction,
   fetchFavoriteFilmsAction,
@@ -12,10 +12,10 @@ import {
 import {ToastContainer} from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
-store.dispatch(fetchPromoFilmAction());
 store.dispatch(fetchFilmsAction());
-store.dispatch(checkAuthAction());
+store.dispatch(fetchPromoFilmAction());
 store.dispatch(fetchFavoriteFilmsAction());
+store.dispatch(checkAuthAction());
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
