+++ project/src/pages/add-review-screen/add-review-screen.tsx
@@ -3,7 +3,7 @@ import {Link, useParams} from 'react-router-dom';
 import Logo from '../../components/logo/logo';
 import ReviewForm from '../../components/review-form/review-form';
 import UserBlock from '../../components/user-block/user-block';
-import {AppRoute, INITIAL_FILM_ID} from '../../const';
+import {AppRoute} from '../../const';
 import {useAppDispatch, useAppSelector} from '../../hooks';
 import {fetchFilmAction} from '../../services/api-actions';
 import {getFilm} from '../../store/app-data/selectors';
@@ -24,7 +24,7 @@ function AddReviewScreen(): JSX.Element {
     dispatch(fetchFilmAction(filmId));
   }, [dispatch, filmId]);
 
-  if (film.id === INITIAL_FILM_ID) {
+  if (filmId !== film.id) {
     return <LoadingScreen />;
   }
