+++ project/src/pages/sign-in-screen/sign-in-screen.tsx
@@ -4,7 +4,7 @@ import Footer from '../../components/footer/footer';
 import Logo from '../../components/logo/logo';
 import {AppRoute, AuthorizationStatus} from '../../const';
 import {useAppDispatch, useAppSelector} from '../../hooks';
-import { useLoginValidation } from '../../hooks/use-login-validation';
+import {useLoginValidation} from '../../hooks/use-login-validation';
 import {loginAction} from '../../services/api-actions';
 import {getAuthorizationStatus} from '../../store/user-process/selectors';
 import {AuthData} from '../../types/auth-data';
