+++ project/src/types/film.ts
@@ -18,9 +18,4 @@ export type Film = {
   isFavorite: boolean,
 }
 
-export type FilmDataType = {
-  id: string,
-  setFilmData: (film: Film) => void,
-}
-
 export type Films = Film[]
